requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sentence-transformers
faiss-cpu
numpy
google-generativeai
langdetect
ollama